refactor: relocate RainbowKit config and add env validation
Addresses reviewer feedback: move config file, remove tsx extension, and implement Zod environment validation
