+++ environment/client.ts
@@ -1,0 +1,22 @@
+// environment/client.ts
+import * as zod from 'zod';
+import { createEnv } from '@t3-oss/env-nextjs';
+
+export const { NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID } = createEnv({
+  client: {
+    NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID: zod.string(),
+  },
+  server: {},
+  runtimeEnv: {
+    NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID: process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID,
+  },
+  emptyStringAsUndefined: true,
+  onValidationError: (issues) => {
+    console.error('Client environment validation errors:', issues);
+    throw new Error(
+      process.env.NODE_ENV === 'production'
+        ? 'Invalid client environment variables'
+        : 'Continuing with missing client environment variables in development'
+    );
+  },
+});
